+++ lib/Model/LineItemData.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * BillingResponseLineItemAllOf
+ * LineItemData
  *
  * PHP version 7.3
  *
@@ -27,7 +27,7 @@ use \ArrayAccess;
 use \Fastly\ObjectSerializer;
 
 /**
- * BillingResponseLineItemAllOf Class Doc Comment
+ * LineItemData Class Doc Comment
  *
  * @category Class
  * @package  Fastly
@@ -36,7 +36,7 @@ use \Fastly\ObjectSerializer;
  * @template TKey int|null
  * @template TValue mixed|null
  */
-class BillingResponseLineItemAllOf implements ModelInterface, ArrayAccess, \JsonSerializable
+class LineItemData implements ModelInterface, ArrayAccess, \JsonSerializable
 {
     public const DISCRIMINATOR = null;
 
@@ -45,7 +45,7 @@ class BillingResponseLineItemAllOf implements ModelInterface, ArrayAccess, \Json
       *
       * @var string
       */
-    protected static $fastlyModelName = 'billing_response_line_item_allOf';
+    protected static $fastlyModelName = 'line_item_data';
 
     /**
       * Array of property to type mappings. Used for (de)serialization
@@ -54,7 +54,7 @@ class BillingResponseLineItemAllOf implements ModelInterface, ArrayAccess, \Json
       */
     protected static $fastlyTypes = [
         'amount' => 'float',
-        'aria_invoice_id' => 'string',
+        'aria_invoice_id' => '\Fastly\Model\LineItemDataReadOnlyInvoiceId',
         'client_service_id' => 'string',
         'credit_coupon_code' => 'string',
         'description' => 'string',
@@ -322,7 +322,7 @@ class BillingResponseLineItemAllOf implements ModelInterface, ArrayAccess, \Json
     /**
      * Gets aria_invoice_id
      *
-     * @return string|null
+     * @return \Fastly\Model\LineItemDataReadOnlyInvoiceId|null
      */
     public function getAriaInvoiceId()
     {
@@ -332,7 +332,7 @@ class BillingResponseLineItemAllOf implements ModelInterface, ArrayAccess, \Json
     /**
      * Sets aria_invoice_id
      *
-     * @param string|null $aria_invoice_id aria_invoice_id
+     * @param \Fastly\Model\LineItemDataReadOnlyInvoiceId|null $aria_invoice_id aria_invoice_id
      *
      * @return self
      */